https://jira.catena-x.net/browse/CATART0PO6-35 move digitaltwins to own folder
