+++ portal/code/tractus-x-portal/src/components/digitaltwins/digitaltwins.tsx
